// frontend/src/app/api-gen/models/migration-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { MigrationTableInfo } from '../models/migration-table-info';
export interface MigrationResponse {
  dry_run: boolean;
  env_a: string;
  env_b: string;
  err_count?: number;
  notes?: Array<string>;
  ok_count?: number;
  tables?: Array<MigrationTableInfo>;
}